[timezone.xml] Update and complete to IANA 2019c
The "timezone.xml" file is not a intended to be a user configurable file.  It is a database of timezones that are not up for arbitrary modifications.  There is a tedious and official process that must be used to have changes made to the official timezone database.

The old data contained duplicate zones that can cause ConfigSelection to fail.  The new version that I have curated is based on the latest timezone data as supplied by IANA (https://www.iana.org/time-zones) and organised by TimezonesDB (https://timezonedb.com/).

This new data is significantly more comprehensive and should have valid timezones for any location in the world.