Update Netherlands
I updates a few providers (which I already posted on Sat4all):
f.e. Delta and some Caiway networks
The Ziggo network should have been standardized, but I left the old networks if it should not be the case everywhere.
Added the new standard Ziggo list and a new provider Kabeltex at the bottem sbecause of the remark "PLEASE DONT SPLIT THE LIST WITH NEW ENTRY'S".